requirements.txt
requests
json
pandas
openai
google-cloud-storage
configparser